updating service section and starting projects section (#26)
